fix: horizontal line interval fixes and refactored
